Pruned ontologies
Deleted Living_Things ontology as it is now included in People. Got rid of the auto generated metadata from People Ontology. It could be confusing when explaining ontology at the Turtle level, especially as some entities didn't have it. Plan to discuss adding metadata automatically but not with People Ontology.